+++ tests/protzilla/data_preprocessing/test_filter_samples.py
@@ -10,7 +10,7 @@ from protzilla.data_preprocessing.filter_samples import (
     by_proteins_missing_plot,
 )
 from tests.protzilla.data_preprocessing.test_peptide_preprocessing import (
-    assert_peptide_filtering_matches_protein_filtering,
+    peptides_df, assert_peptide_filtering_matches_protein_filtering,
 )
