+++ tests/protzilla/data_preprocessing/test_outlier_detection.py
@@ -11,7 +11,7 @@ from protzilla.data_preprocessing.outlier_detection import (
     by_pca_plot,
 )
 from tests.protzilla.data_preprocessing.test_peptide_preprocessing import (
-    assert_peptide_filtering_matches_protein_filtering,
+    peptides_df, assert_peptide_filtering_matches_protein_filtering,
 )
 
 # TODO #21: implement actual tests for outlier detection
@@ -65,7 +65,8 @@ def outlier_detection_df_with_nan():
 
 
 def test_outlier_detection_with_isolation_forest(
-    show_figures, outlier_detection_df, peptides_df
+    show_figures, outlier_detection_df,
+        peptides_df
 ):
     method_inputs = {
         "protein_df": outlier_detection_df,
