+++ tests/protzilla/data_preprocessing/test_peptide_preprocessing.py
@@ -1,4 +1,5 @@
 import pandas as pd
+import pytest
 
 from protzilla.constants.paths import TEST_DATA_PATH
 from protzilla.data_preprocessing.peptide_filter import by_pep_value, by_pep_value_plot
@@ -56,3 +57,37 @@ def test_pep_filter(show_figures, leftover_peptide_df, filtered_peptides_list):
 
     pd.testing.assert_frame_equal(method_outputs["peptide_df"], leftover_peptide_df)
     assert method_outputs["filtered_peptides"] == filtered_peptides_list
+
+
+@pytest.fixture
+def peptides_df():
+    df = pd.DataFrame(
+        (
+            ["Sample1", "Protein1", "SEQA", 1000000, 0.00001],
+            ["Sample1", "Protein2", "SEQB", 2000000, 0.00002],
+            ["Sample1", "Protein2", "SEQC", 3000000, 0.00003],
+            ["Sample1", "Protein2", "SEQD", 4000000, 0.00004],
+            ["Sample1", "Protein3", "SEQE", 5000000, 0.00005],
+            ["Sample1", "Protein3", "SEQF", 6000000, 0.00006],
+            ["Sample1", "Protein3", "SEQG", 7000000, 0.00007],
+            ["Sample1", "Protein4", "SEQH", 8000000, 0.00008],
+            ["Sample1", "Protein5", "SEQI", 9000000, 0.00009],
+            ["Sample2", "Protein1", "SEQJ", 10000000, 0.0001],
+            ["Sample2", "Protein2", "SEQK", 11000000, 0.00011],
+            ["Sample2", "Protein3", "SEQL", 12000000, 0.00012],
+            ["Sample2", "Protein4", "SEQM", 13000000, 0.00013],
+            ["Sample2", "Protein5", "SEQN", 14000000, 0.00014],
+            ["Sample3", "Protein1", "SEQO", 15000000, 0.00015],
+            ["Sample3", "Protein2", "SEQP", 16000000, 0.00016],
+            ["Sample3", "Protein3", "SEQQ", 17000000, 0.00017],
+            ["Sample3", "Protein4", "SEQR", 18000000, 0.00018],
+            ["Sample3", "Protein5", "SEQS", 19000000, 0.00019],
+            ["Sample4", "Protein1", "SEQT", 20000000, 0.0002],
+            ["Sample4", "Protein2", "SEQU", 21000000, 0.00021],
+            ["Sample4", "Protein3", "SEQV", 22000000, 0.00022],
+            ["Sample4", "Protein4", "SEQW", 23000000, 0.00023],
+        ),
+        columns=["Sample", "Protein ID", "Sequence", "Intensity", "PEP"],
+    )
+
+    return df
